Factor rgb_to_hex, hex_to_int and int_to_hex away from StringMobject and to utils/color
